refactor: refactor pages files and folders
