+++ ecommerce-application/src/pages/Beverages/Beverages.tsx
